Remove unused lambda expressions assigned to "_"
This commit removes several noop lambda expressions assigned to the name
"_". In all but one case, they were unused. In the case where it was
being called, the call was simply removed (the lambda was just returing
the value passed to it, so it was unnecessary).

_('foo') is a common idiom for localization, and my guess is that at
some point some code using that convention was copy/pasted in to this
project, and the lambda was added as a workaround for that function
being missing.