Escape strings from the database when rendering
Disabling autoescaping is unsafe, and results in an XSS vulnerability if
the attacker is able to control database values which get included in
DSE rendered reports.

This commit enables autoescaping by default, and adds a clearly named
setting to restore the old behavior in case someone was relying on it.