Run queries as atomic requests
Without this, a query error (eg. a typo in a table name) causes a 500-level error if the default database connection has ATOMIC_REQUESTS set to True. This is because once the query fails, no other database query succeeds (eg. trying to access the database to render the error page).